feat(tsconfig): refactor tsconfig utilities and enhance parsing functionality
